cava folders structure reorganization
