+++ cavacore_test.c
@@ -2,7 +2,7 @@
 // gcc -c -g cavacore_test.c
 // gcc -o cavacore_test cavacore_test.o build/libcavacore.a -lm -lfftw3
 
-#include "cavacore.h"
+#include "include/cava/cavacore.h"
 #include <math.h>
 #include <stdio.h>
