+++ include/cava/input/pulse.h
