+++ include/cava/output/terminal_bcircle.h
