+++ src/cavacore.c
@@ -1,4 +1,4 @@
-#include "cavacore.h"
+#include "cava/cavacore.h"
 #ifndef M_PI
 #define M_PI 3.1415926535897932385
 #endif
