+++ src/input/oss.c
@@ -4,8 +4,8 @@
 #include <sys/ioctl.h>
 #include <sys/soundcard.h>
 
-#include "input/common.h"
-#include "input/oss.h"
+#include "../../include/cava/input/oss.h"
+#include "cava/input/common.h"
 
 static bool set_format(int fd, struct audio_data *audio) {
     // CAVA favors signed and little endian (sle) formats. It might actually not work correctly if
