+++ src/output/shaders/bar_spectrum.frag
