+++ src/output/shaders/northern_lights.frag
