+++ src/output/shaders/spectrogram.frag
