+++ src/output/shaders/winamp_line_style_spectrum.frag
