+++ src/output/terminal_bcircle.c
@@ -1,4 +1,4 @@
-#include "output/terminal_bcircle.h"
+#include "../../include/cava/output/terminal_bcircle.h"
 
 #include <curses.h>
 #include <locale.h>
