+++ src/output/terminal_ncurses.c
@@ -1,11 +1,10 @@
-#include "output/terminal_ncurses.h"
+#include "../../include/cava/output/terminal_ncurses.h"
+#include "../../include/cava/util.h"
 
 #include <curses.h>
 #include <stdlib.h>
 #include <string.h>
 #include <wchar.h>
-
-#include "util.h"
 
 int gradient_size = 64;
