+++ src/output/themes/tricolor
